% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_LLi.R
\name{exp_LLi}
\alias{exp_LLi}
\title{Likelihood function of the discrete exponential distribution}
\usage{
exp_LLi(k, kmin, params, log = T)
}
\arguments{
\item{k}{a numeric vector of each observed degree}

\item{kmin}{right-tail cutoff that should be estimated via the \code{igraph} package with the \code{fit_power_law(...)} function}

\item{params}{a list object holding the lambda parameter for the discrete exponential distribution}

\item{log}{a logical scalar. Defaults to TRUE which tells the function to return the log-likelihood instead of the likelihood.}
}
\value{
a numeric vector with length equal to the number of data point that is greater than or equal to \code{kmin}.
Each element is the likelihood value of the data evaluated at the given parameter \code{lambda} and right-tail cutoff \code{kmin}.
}
\description{
Computes the likelihood of the discrete exponential model with a given \code{kmin}
}
\note{
Note that the \code{kmin} is inclusive, as in, for a vector \code{k = c(2,1,3,5,4)}, if \code{kmin} is 3, then the data analyzed would be \code{c(3, 5, 4)}.
}
\examples{
# a data vector c(1,2,3,4,5) with kmin = 1, thus all data are analyzed
exp_LLi(k = c(1,2,3,4,5), kmin = 1, params = list(lambda = 2.5))

# a data vector c(2,1,3,5,4) with kmin = 3, thus only c(3, 5, 4) are analyzed
# note that shifting the kmin changes the location of the distribution,
# hence k=3 here will not yield the same likelihood value compare to the previous example
exp_LLi(k = c(2,1,3,5,4), kmin = 3, params = list(lambda = 2.5))
}
