% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lognorm.R
\name{fit_lognorm}
\alias{fit_lognorm}
\title{Fits the Discretized Log-Normal Model}
\usage{
fit_lognorm(
  k,
  kmin,
  inits = c(mu = 0, log.sigma = 1),
  opt.controls = list(fnscale = -sum(k >= kmin), reltol = 1e-16, maxit = 1e+05)
)
}
\arguments{
\item{k}{a numeric vector of each observed degree}

\item{kmin}{a numeric scalar to indicate where the right-tail cutoff is}

\item{inits}{a named vector indicating the initial values for the numerical optimization}

\item{opt.controls}{control parameter for the optimization routine}
}
\value{
a list object containing the following elements:
\code{par} is the parameter estimate in their original scale (e.g., labmda > 0). \code{log.par} is the estimated parameter from the \code{optim(...)} scaled to be ranged from (-Inf, Inf).
\code{LL} is the overall log-likelihood value evaluated at the estimated MLE, \code{optim} is the output from \code{optim(...)}.
\code{data} is a numeric vector for the data that was analyzed (i.e., where k >= kmin), \code{kmin} is the user specified \code{kmin}, and \code{data.orig} is the original vector of input data
}
\description{
Standard numerical optimization routine to obtain the maximum likelihood estimate for the discrete log-normal model.
}
\note{
For the discrete exponential model, the Nelder-Mead method is used for numerical optimization
}
\examples{
# a data vector
set.seed(1)
my_k = rpois(n = 100, lambda = 2)
my_kmin = 2

# fit
my_opt = fit_lognorm(
  k = my_k, kmin = my_kmin,
  inits = c(mu = 0, log.sigma = 1)
)
}
