% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_PL.R
\name{gof_PL}
\alias{gof_PL}
\title{Semiparametric GoF test for the power law model}
\usage{
gof_PL(k, kmin, alpha, Nsim = 100, ss)
}
\arguments{
\item{k}{a numeric vector for the observed data}

\item{kmin}{a numeric scalar indicating the right-tail cutoff}

\item{alpha}{a numeric scalar for the estimated \code{alpha} parameter from the power law model (can be obtained via \code{igraph::fit_power_law})}

\item{Nsim}{number of times the simulations should be ran to obtain theoretical KS value (the more simulation, the more accurate the result).
For accuracy up to 0.01 decimal point, use \code{Nsim = 2500}. See Clauset's paper:
\verb{Clauset, A., Shalizi, C. R., & Newman, M. E. (2009). Power-law distributions in empirical data. SIAM review, 51(4), 661-703.}}

\item{ss}{an integer value for setting simulation seed.}
}
\value{
a matrix object with number of rows equal to \code{Nsim}, where each row stores the simulated statistic obtained by fitting the power law model to the synthetic data
}
\description{
Performs the sempiparametric approach to goodness of fit test described in \verb{https://arxiv.org/pdf/0706.1062.pdf} (pg 17; paragraph 2-3).
}
\examples{
# generate some dataset
set.seed(1)
my_N = 500
my_k = poweRlaw::rpldis(n = my_N, xmin = 3, alpha = 5)

# power law fit
my_PLmodel = igraph::fit_power_law(x = my_k)

# goodness of fit
ss = gof_PL(k = my_k, kmin = my_PLmodel$xmin, my_PLmodel$alpha, Nsim = 1000, ss = 123)
mean(ss[, "KS.stat"] > my_PLmodel$KS.stat)
}
