% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTest.R
\name{LRTest}
\alias{LRTest}
\title{Likelihood Ratio test to compare model performance}
\usage{
LRTest(LLi1, LLi2)
}
\arguments{
\item{LLi1}{a numeric vector of the log-likelihood of a fitted model}

\item{LLi2}{a numeric vector of the log-likelihood of an other fitted model}
}
\value{
a list object containing the test statistics "R" and p-value (among some other relevant statistics).
\code{Rstat} is the R statistic, use its sign to tell which model performed better. \code{Rstat.sd} is the SE of the R statistic
\code{pval} is the test p-value, \code{L1} is the overall log-likelihood of model 1, and \code{L2} is the overall log-likelihood value for model 2
}
\description{
Performs the likelihood ratio test as described in \verb{https://www.nature.com/articles/s41467-019-08746-5?ref=https://githubhelp.com}
}
\note{
This function takes LLi1 minus LLi2 to obtain the R statistic.
Hence a positive R statistic may indicate model 1 is better than model 2 (if p-value agrees). Simply running a paired t-test on LLi1 and LLi2 usually yield the same result
}
\examples{
# generate data and some theoretical likelihood
set.seed(1)
my_data = rpois(n = 100, lambda = 5)
my_data.PL = PL_LLi(k = my_data, kmin = 4, params = list(alpha = 2.1))
my_data.weib = weib_LLi(k = my_data, kmin = 4, params = list(a = 1.5, b = 1.5))

# perform LRT
LRTest(LLi1 = my_data.PL, LLi2 = my_data.weib) # R statistic is greater than 0, p-value is < 0.05. So we can conclude that the power law model in this case was a better fit than Weibull
}
